smart-audit-backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
web3
py-solc-x
python-multipart
pymongo